Add posts as a feature module
