--- app/src/main/java/com/addhen/checkin/view/widget/SquaredFrameLayout.kt
@@ -1,36 +1,0 @@
-/*
- * Copyright (c) 2015 - 2018 Henry Addo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package com.addhen.checkin.view.widget
-
-import android.content.Context
-import android.util.AttributeSet
-import android.widget.FrameLayout
-
-class SquaredFrameLayout : FrameLayout {
-
-  constructor(context: Context) : super(context)
-
-  constructor(context: Context, attrs: AttributeSet) : super(context, attrs)
-
-  constructor(context: Context, attrs: AttributeSet, defStyleAttr: Int) : super(context, attrs,
-      defStyleAttr)
-
-  override fun onMeasure(widthMeasureSpec: Int, heightMeasureSpec: Int) {
-    super.onMeasure(widthMeasureSpec, widthMeasureSpec)
-  }
-}
