+++ app/src/main/res/values/strings.xml
@@ -1,21 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
   <!-- App -->
-  <string name="retry">Retry</string>
-  <string name="save">Save</string>
-  <string name="okay">Okay</string>
-  <string name="share">Share</string>
-  <string name="like">Like</string>
-  <!-- Posts -->
-  <string name="posts_load_failed">Failed to load posts.</string>
-  <string name="posts">Posts</string>
-  <string name="post_details">Post Details</string>
-  <string name="add_post">Add Post</string>
-  <plurals name="likes_count">
-    <item quantity="one">%d like</item>
-    <item quantity="other">%d likes</item>
-  </plurals>
-  <string name="empty_post_list">No posts yet.</string>
   <string name="no_comments">No comments.</string>
   <string name="comment">Comment</string>
   <string name="main_menu_search">Search</string>
@@ -28,11 +13,6 @@
     the action again.
   </string>
   <string name="sign_in_cancelled">Sign in cancelled</string>
-  <!-- User profile -->
-  <string name="user_profile">User Profile</string>
-  <string name="user_profile_signout">Signout</string>
-  <string name="following">Following</string>
-  <string name="followers">Followers</string>
   <!-- Comments -->
   <string name="comments_load_failed">Failed to load comments.</string>
   <plurals name="comments_count">
