--- app/src/main/res/values/styles_posts.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-  <style name="PostMetaTextCaption">
-    <item name="android:layout_width">wrap_content</item>
-    <item name="android:layout_height">wrap_content</item>
-    <item name="android:textAppearance">@style/TextCaption</item>
-    <item name="android:textColor">@color/grey600</item>
-    <item name="android:ellipsize">end</item>
-    <item name="android:focusable">true</item>
-    <item name="android:focusableInTouchMode">true</item>
-    <item name="android:lines">1</item>
-  </style>
-
-  <style name="PostIconText" parent="PostMetaTextCaption">
-    <item name="android:textSize">@dimen/text_14sp</item>
-  </style>
-
-  <style name="PostTitle">
-    <item name="android:textAppearance">@style/TextCaption</item>
-  </style>
-</resources>
