+++ base/src/main/res/drawable/ic_add_white_24dp.xml
