+++ base/src/main/res/drawable/ic_circular_avatar_mask.xml
