+++ base/src/main/res/drawable/ic_favorite_border_grey_24dp.xml
