+++ base/src/main/res/values/strings.xml
@@ -2,4 +2,14 @@
 <resources>
   <string name="app_name">CheckIn</string>
   <string name="logo_description">App logo</string>
+  <!-- User profile -->
+  <string name="user_profile">User Profile</string>
+  <string name="user_profile_signout">Signout</string>
+  <string name="following">Following</string>
+  <string name="followers">Followers</string>
+  <string name="retry">Retry</string>
+  <string name="save">Save</string>
+  <string name="okay">Okay</string>
+  <string name="share">Share</string>
+  <string name="like">Like</string>
 </resources>
