+++ base/src/main/res/values/styles.xml
@@ -99,4 +99,15 @@
     <item name="android:layout_gravity">center</item>
     <item name="android:textAppearance">@style/TextEmpty</item>
   </style>
+
+  <style name="MetaTextCaption">
+    <item name="android:layout_width">wrap_content</item>
+    <item name="android:layout_height">wrap_content</item>
+    <item name="android:textAppearance">@style/TextCaption</item>
+    <item name="android:textColor">@color/grey600</item>
+    <item name="android:ellipsize">end</item>
+    <item name="android:focusable">true</item>
+    <item name="android:focusableInTouchMode">true</item>
+    <item name="android:lines">1</item>
+  </style>
 </resources>
