+++ build.gradle.kts
@@ -13,7 +13,7 @@ buildscript {
     classpath(kotlin("gradle-plugin", version = Versions.KOTLIN))
     classpath("com.google.gms:google-services:3.2.0")
     // NOTE: Do not place your application dependencies here; they belong
-    // in the individual module build.gradle files
+    // in the individual module build.gradle.kts files
   }
 }
