+++ posts/src/main/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<manifest package="com.addhen.checkin.posts"/>
+
