+++ posts/src/main/java/com/addhen/checkin/posts/view/PostsBuilder.kt
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2015 - 2018 Henry Addo
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package com.addhen.checkin.posts.view
+
+import androidx.lifecycle.ViewModel
+import com.addhen.checkin.base.di.module.ViewModelKey
+import dagger.Binds
+import dagger.Module
+import dagger.android.ContributesAndroidInjector
+import dagger.multibindings.IntoMap
+
+@Module
+abstract class PostsBuilder {
+  @ContributesAndroidInjector
+  abstract fun postsFragment(): PostsFragment
+
+  @Binds
+  @IntoMap
+  @ViewModelKey(PostsViewModel::class)
+  abstract fun postsViewModel(viewModel: PostsViewModel): ViewModel
+}
