+++ posts/src/main/res/anim/slide_out_likes_counter.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android"
+     android:duration="150"
+     android:interpolator="@android:anim/linear_interpolator">
+  <translate
+      android:fromYDelta="0"
+      android:toYDelta="-80%p"/>
+  <alpha
+      android:fromAlpha="1.0"
+      android:toAlpha="0.0"/>
+</set>
