+++ posts/src/main/res/layout/post_item.xml
@@ -1,0 +1,171 @@
+<?xml version="1.0" encoding="utf-8"?><!--
+  ~ Copyright (c) 2015 - 2018 Henry Addo
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~     http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  -->
+
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:app="http://schemas.android.com/apk/res-auto">
+  <data>
+    <variable
+        name="viewModel"
+        type="com.addhen.checkin.posts.view.PostItemViewModel"/>
+  </data>
+
+  <androidx.cardview.widget.CardView
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:layout_margin="@dimen/space_8dp"
+      android:background="?attr/selectableItemBackground"
+      android:clickable="true"
+      android:focusable="true"
+      android:onClick="@{(view) -> viewModel.onClickPost()}"
+      app:cardCornerRadius="@dimen/cardview_default_radius">
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical">
+
+      <RelativeLayout
+          android:layout_width="match_parent"
+          android:layout_height="wrap_content"
+          android:layout_margin="@dimen/space_8dp">
+
+        <ImageView
+            android:id="@+id/user_profile"
+            android:layout_width="@dimen/user_avatar_40dp"
+            android:layout_height="@dimen/user_avatar_40dp"
+            android:layout_alignParentStart="true"
+            android:layout_alignParentTop="true"
+            android:contentDescription="@string/user_profile"
+            app:srcCompat="@drawable/ic_circular_avatar_mask"
+            app:userImageSize="@{@dimen/user_avatar_40dp}"
+            app:userImageUrl="@{viewModel.post.user.photoUrl}"/>
+
+        <TextView
+            android:id="@+id/user_name"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_alignParentTop="true"
+            android:layout_marginBottom="@dimen/space_8dp"
+            android:layout_marginStart="@dimen/space_8dp"
+            android:layout_toEndOf="@id/user_profile"
+            android:layout_toStartOf="@id/post_timestamp"
+            android:textAppearance="@style/TextSubheading"
+            app:userName="@{viewModel.post.user}"/>
+
+        <TextView
+            android:id="@+id/post_timestamp"
+            style="@style/MetaTextCaption"
+            android:layout_alignParentEnd="true"
+            android:layout_alignParentTop="true"
+            android:layout_marginBottom="@dimen/space_8dp"
+            android:drawablePadding="@dimen/icon_space_4dp"
+            android:drawableStart="@drawable/state_access_time"
+            android:gravity="center|center_horizontal|center_vertical"
+            app:timeAgo="@{viewModel.post.created}"/>
+
+        <TextView
+            android:id="@+id/post_place_name"
+            style="@style/MetaTextCaption"
+            android:layout_below="@id/user_name"
+            android:layout_marginStart="@dimen/space_8dp"
+            android:layout_toEndOf="@id/user_profile"
+            android:layout_toStartOf="@id/post_timestamp"
+            android:text="@{viewModel.post.location.name}"/>
+
+        <TextView
+            android:id="@+id/post_city_country"
+            style="@style/MetaTextCaption"
+            android:layout_alignParentEnd="true"
+            android:layout_below="@id/post_timestamp"
+            android:layout_toEndOf="@id/post_place_name"
+            android:gravity="end"
+            android:text="@{viewModel.post.location.country}"/>
+      </RelativeLayout>
+
+
+      <TextView
+          android:id="@+id/post_content"
+          android:layout_width="match_parent"
+          android:layout_height="wrap_content"
+          android:layout_margin="@dimen/space_8dp"
+          android:text="@{viewModel.post.content}"
+          android:textAppearance="@style/TextBody1"/>
+
+      <ImageView
+          android:id="@+id/posts_bottom"
+          android:layout_width="match_parent"
+          android:layout_height="wrap_content"
+          android:contentDescription="@string/posts"/>
+
+      <LinearLayout
+          android:layout_width="match_parent"
+          android:layout_height="wrap_content"
+          android:layout_margin="@dimen/space_8dp">
+
+        <ImageButton
+            android:id="@+id/post_like_btn"
+            android:layout_width="@dimen/social_icons"
+            android:layout_height="@dimen/social_icons"
+            android:contentDescription="@string/like"
+            android:background="@android:color/transparent"
+            app:srcCompat="@drawable/ic_favorite_border_grey_24dp"/>
+
+        <Button
+            android:id="@+id/post_comment_btn"
+            style="@style/MetaTextCaption.IconText"
+            android:layout_width="@dimen/social_icons"
+            android:layout_height="wrap_content"
+            android:background="?selectableItemBackgroundBorderless"
+            android:drawablePadding="@dimen/icon_space_4dp"
+            android:drawableStart="@drawable/ic_comment"
+            android:text="@{``+33}"/>
+
+        <LinearLayout
+            android:layout_width="0dp"
+            android:layout_height="match_parent"
+            android:layout_weight="1"
+            android:gravity="center_vertical|end">
+
+          <ImageView
+              android:layout_width="wrap_content"
+              android:layout_height="wrap_content"
+              android:contentDescription="@string/like"
+              app:srcCompat="@drawable/ic_favorite_red_24dp"/>
+
+          <TextSwitcher
+              android:id="@+id/post_like_counter"
+              android:layout_width="wrap_content"
+              android:layout_height="wrap_content"
+              android:inAnimation="@anim/slide_in_likes_counter"
+              android:outAnimation="@anim/slide_out_likes_counter"
+              android:text="@{@plurals/likes_count((int) viewModel.post.likes, (int) viewModel.post.likes)}">
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:textColor="@color/rouge_54"/>
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:textColor="@color/rouge_54"/>
+          </TextSwitcher>
+        </LinearLayout>
+      </LinearLayout>
+    </LinearLayout>
+  </androidx.cardview.widget.CardView>
+</layout>
