+++ posts/src/main/res/layout/posts_fragment.xml
@@ -1,0 +1,73 @@
+<?xml version="1.0" encoding="utf-8"?><!--
+  ~ Copyright (c) 2015 - 2018 Henry Addo
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~     http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  ~
+  -->
+
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:app="http://schemas.android.com/apk/res-auto">
+
+  <data>
+
+    <variable
+        name="viewModel"
+        type="com.addhen.checkin.posts.view.PostsViewModel"/>
+  </data>
+
+  <androidx.coordinatorlayout.widget.CoordinatorLayout
+      android:layout_width="match_parent"
+      android:layout_height="match_parent">
+
+    <FrameLayout
+        android:id="@+id/post_root_framelayout"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_marginBottom="@dimen/space_54dp">
+
+      <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
+          android:id="@+id/swipe_refresh_layout"
+          android:layout_width="match_parent"
+          android:layout_height="match_parent"
+          app:onRefreshListener="@{viewModel::onSwipeRefresh}">
+
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/posts_recycler_view"
+            style="@style/BaseRecyclerView"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"/>
+      </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
+
+      <ProgressBar
+          android:id="@+id/loading_progress_bar"
+          android:layout_width="wrap_content"
+          android:layout_height="wrap_content"
+          android:layout_gravity="center"
+          android:indeterminate="true"
+          android:visibility="gone"/>
+
+      <TextView
+          android:id="@+id/empty_view_header"
+          style="@style/BaseEmptyView"
+          android:text="@string/empty_post_list"
+          android:visibility="gone"/>
+
+      <com.google.android.material.floatingactionbutton.FloatingActionButton
+          android:id="@+id/fab"
+          style="@style/BaseFabView"
+          android:layout_width="wrap_content"
+          app:pressedTranslationZ="@dimen/translation_z_12dp"
+          app:srcCompat="@drawable/ic_add_white_24dp"/>
+    </FrameLayout>
+  </androidx.coordinatorlayout.widget.CoordinatorLayout>
+</layout>
